26.05 Código Comentado

+++ src/user/user.service.ts
@@ -47,21 +47,26 @@ export class UserService {
     return users;
   }
 
+  // Pega a imagem de um usuário
   async getImage(id: number) {
     const user = await this.prisma.user.findUnique({
         where: {
             id: id
         }
     })
+    // Se o usuário não existir retorna um erro
     if (!user) throw new HttpException('Usuário inválido', HttpStatus.BAD_REQUEST)
 
+    // Procura a imagem
     const [base, finalArquivo] = user.imgPath.split('.')
     let filePath = path.join(process.cwd(), `uploads/images/Photo-${id}.${finalArquivo}`)
 
+    // Verifica se existe uma imagem própria, se não coloca uma padrão
     if (!fs.existsSync(filePath)) {
       filePath = path.join(process.cwd(), `uploads/fixo/avatar.png`);
     }
 
+    // Retorna o caminho da imagem
     return filePath;
   }
 
@@ -138,6 +143,7 @@ export class UserService {
     return "Usuário deletado"
   }
 
+  // Salva a imagem e 
   async newImage(id: number, file: Express.Multer.File) {
     const user = await this.prisma.user.findUnique({
       where: {
@@ -145,19 +151,30 @@ export class UserService {
       }
     })
 
+    // Se o usuário não existir retorna um erro
     if (!user) throw new HttpException('Usuário não existe', HttpStatus.BAD_REQUEST);
 
+    // Extensões válidas e extensão do arquivo   
+    const allowedExtensions = ['.jpeg', '.jpg', '.png'];
+    const fileExt = path.extname(file.originalname);
+    
+    // Se o arquivo tiver outras extensões diz que é inválido 
+    if (!allowedExtensions.includes(fileExt)) throw new HttpException('Extensão de arquivo não permitida.', HttpStatus.BAD_REQUEST);
+
+    // Verifica se já existe uma imagem salva
     if (user.imgPath) {
       const imagePath = path.join(process.cwd(), 'uploads', user.imgPath); 
 
+      // Verifica se existe no local e se existir deleta
       if (fs.existsSync(imagePath)) {
         fs.unlinkSync(imagePath);
       } 
     }
 
-    const fileExt = path.extname(file.originalname);
+    // Configura o nome do arquivo 
     const fileName = `Photo-${id}${fileExt}`;
 
+    // Atualiza na database
     await this.prisma.user.update({
       where: {
         id: id
@@ -167,16 +184,14 @@ export class UserService {
       }
     })
 
-    const allowedExtensions = ['.jpeg', '.jpg', '.png'];
-
-    if (!allowedExtensions.includes(fileExt)) throw new HttpException('Extensão de arquivo não permitida.', HttpStatus.BAD_REQUEST);
-
+    // Acha o caminho do arquivo e salva ele
     const destination = path.join(process.cwd(), 'uploads/image');
     const filePath = `${destination}\\${fileName}`;
     fs.writeFile(filePath, file.buffer, (err) => {
       if (err) throw new HttpException('Erro interno do servidor', HttpStatus.INTERNAL_SERVER_ERROR);
     })
 
+    // Retorna o caminho do arquivo
     return filePath;
   }
 }
